updated sale to not display the drugs which has got amount less than 0 or equal to 0

+++ Core/Entities/Druggist.cs
@@ -9,7 +9,7 @@ namespace Core.Entities
         public string Name { get; set; }
         public string Surname { get; set; }
         public byte Age { get; set; }
-        public uint Experience { get; set; }
+        public double Experience { get; set; }
         public DrugStore DrugStore { get; set; }
     }
 }
